ARCO-155: separate tx validators
